Move `http-body-0-4-x` specific code to separate modules
