+++ rust-runtime/aws-smithy-types/src/byte_stream.rs
@@ -143,6 +143,12 @@ pub mod error;
 #[cfg(feature = "rt-tokio")]
 pub use self::bytestream_util::FsBuilder;
 
+/// This module is named after the `http-body` version number since we anticipate
+/// needing to provide equivalent functionality for 1.x of that crate in the future.
+/// The name has a suffix `_x` to avoid name collision with a third-party `http-body-0-4`.
+#[cfg(feature = "http-body-0-4-x")]
+pub mod http_body_0_4_x;
+
 pin_project! {
     /// Stream of binary data
     ///
@@ -229,11 +235,14 @@ pin_project! {
     ///     ```
     ///
     /// 2. **From a file**: ByteStreams created from a path can be retried. A new file descriptor will be opened if a retry occurs.
+    ///
+    ///     _Note: The `http-body-0-4-x` feature must be active to call `ByteStream::from_body_0_4`._
+    ///
     ///     ```no_run
-    ///     #[cfg(feature = "tokio-rt")]
+    ///     #[cfg(all(feature = "tokio-rt", feature = "http-body-0-4-x"))]
     ///     # {
     ///     use aws_smithy_types::byte_stream::ByteStream;
-    ///     let stream = ByteStream::from_path("big_file.csv");
+    ///     let stream = ByteStream::from_path_body_0_4("big_file.csv");
     ///     # }
     ///     ```
     ///
@@ -281,18 +290,6 @@ impl ByteStream {
         }
     }
 
-    #[cfg(feature = "http-body-0-4-x")]
-    /// Construct a `ByteStream` from a type that implements [`http_body_0_4::Body<Data = Bytes>`](http_body_0_4::Body).
-    ///
-    /// _Note: This is only available with `http-body-0-4-x` enabled._
-    pub fn from_body_0_4<T, E>(body: T) -> Self
-    where
-        T: http_body_0_4::Body<Data = Bytes, Error = E> + Send + Sync + 'static,
-        E: Into<crate::body::Error> + 'static,
-    {
-        ByteStream::new(SdkBody::from_body_0_4(body))
-    }
-
     /// Consume the `ByteStream`, returning the wrapped SdkBody.
     // Backwards compatibility note: Because SdkBody has a dyn variant,
     // we will always be able to implement this method, even if we stop using
@@ -312,6 +309,9 @@ impl ByteStream {
     #[cfg(feature = "byte-stream-poll-next")]
     /// Attempt to pull out the next value of this stream, returning `None` if the stream is
     /// exhausted.
+    // This should only be used when one needs to implement a trait method like
+    // `futures_core::stream::Stream::poll_next` on a new-type wrapping a `ByteStream`.
+    // In general, use the `next` method instead.
     pub fn poll_next(
         self: Pin<&mut Self>,
         cx: &mut Context<'_>,
@@ -351,83 +351,6 @@ impl ByteStream {
     /// ```
     pub async fn collect(self) -> Result<AggregatedBytes, Error> {
         self.inner.collect().await.map_err(Error::streaming)
-    }
-
-    /// Returns a [`FsBuilder`](crate::byte_stream::FsBuilder), allowing you to build a `ByteStream` with
-    /// full control over how the file is read (eg. specifying the length of the file or the size of the buffer used to read the file).
-    /// ```no_run
-    /// # #[cfg(all(feature = "rt-tokio", feature = "http-body-0-4-x"))]
-    /// # {
-    /// use aws_smithy_types::byte_stream::{ByteStream, Length};
-    ///
-    /// async fn bytestream_from_file() -> ByteStream {
-    ///     let bytestream = ByteStream::read_with_body_0_4_from()
-    ///         .path("docs/some-large-file.csv")
-    ///         // Specify the size of the buffer used to read the file (in bytes, default is 4096)
-    ///         .buffer_size(32_784)
-    ///         // Specify the length of the file used (skips an additional call to retrieve the size)
-    ///         .length(Length::Exact(123_456))
-    ///         .build()
-    ///         .await
-    ///         .expect("valid path");
-    ///     bytestream
-    /// }
-    /// # }
-    /// ```
-    #[cfg(all(feature = "rt-tokio", feature = "http-body-0-4-x"))]
-    #[cfg_attr(docsrs, doc(cfg(feature = "rt-tokio")))]
-    pub fn read_with_body_0_4_from() -> FsBuilder {
-        FsBuilder::new_with_body_0_4()
-    }
-
-    /// Create a ByteStream that streams data from the filesystem
-    ///
-    /// This function creates a retryable ByteStream for a given `path`. The returned ByteStream
-    /// will provide a size hint when used as an HTTP body. If the request fails, the read will
-    /// begin again by reloading the file handle.
-    ///
-    /// ## Warning
-    /// The contents of the file MUST not change during retries. The length & checksum of the file
-    /// will be cached. If the contents of the file change, the operation will almost certainly fail.
-    ///
-    /// Furthermore, a partial write MAY seek in the file and resume from the previous location.
-    ///
-    /// Note: If you want more control, such as specifying the size of the buffer used to read the file
-    /// or the length of the file, use a [`FsBuilder`](crate::byte_stream::FsBuilder) as returned
-    /// from `ByteStream::read_with_body_0_4_from`
-    ///
-    /// # Examples
-    /// ```no_run
-    /// use aws_smithy_types::byte_stream::ByteStream;
-    /// use std::path::Path;
-    ///  async fn make_bytestream() -> ByteStream {
-    ///     ByteStream::from_path_body_0_4("docs/rows.csv").await.expect("file should be readable")
-    /// }
-    /// ```
-    #[cfg(all(feature = "rt-tokio", feature = "http-body-0-4-x"))]
-    #[cfg_attr(
-        docsrs,
-        doc(cfg(all(feature = "rt-tokio", feature = "http-body-0-4-x")))
-    )]
-    pub async fn from_path_body_0_4(path: impl AsRef<std::path::Path>) -> Result<Self, Error> {
-        FsBuilder::new_with_body_0_4().path(path).build().await
-    }
-
-    /// Create a ByteStream from a file
-    ///
-    /// NOTE: This will NOT result in a retryable ByteStream. For a ByteStream that can be retried in the case of
-    /// upstream failures, use [`ByteStream::from_path_body_0_4`](ByteStream::from_path_body_0_4)
-    #[deprecated(
-        since = "0.40.0",
-        note = "Prefer the more extensible ByteStream::read_from() API"
-    )]
-    #[cfg(all(feature = "rt-tokio", feature = "http-body-0-4-x"))]
-    #[cfg_attr(
-        docsrs,
-        doc(cfg(all(feature = "rt-tokio", feature = "http-body-0-4-x")))
-    )]
-    pub async fn from_file_body_0_4(file: tokio::fs::File) -> Result<Self, Error> {
-        FsBuilder::new_with_body_0_4().file(file).build().await
     }
 
     #[cfg(feature = "rt-tokio")]
@@ -503,13 +426,6 @@ impl From<Bytes> for ByteStream {
 impl From<Vec<u8>> for ByteStream {
     fn from(input: Vec<u8>) -> Self {
         Self::from(Bytes::from(input))
-    }
-}
-
-#[cfg(all(feature = "http-body-0-4-x", feature = "hyper-0-14-x"))]
-impl From<hyper_0_14::Body> for ByteStream {
-    fn from(input: hyper_0_14::Body) -> Self {
-        ByteStream::new(SdkBody::from_body_0_4(input))
     }
 }
 
@@ -627,63 +543,6 @@ mod tests {
                 .into_bytes(),
             Bytes::from("a simple body")
         );
-    }
-
-    #[cfg(feature = "http-body-0-4-x")]
-    #[tokio::test]
-    async fn read_from_channel_body() {
-        let (mut sender, body) = hyper_0_14::Body::channel();
-        let byte_stream = Inner::new(SdkBody::from_body_0_4(body));
-        tokio::spawn(async move {
-            sender.send_data(Bytes::from("data 1")).await.unwrap();
-            sender.send_data(Bytes::from("data 2")).await.unwrap();
-            sender.send_data(Bytes::from("data 3")).await.unwrap();
-        });
-        assert_eq!(
-            byte_stream.collect().await.expect("no errors").into_bytes(),
-            Bytes::from("data 1data 2data 3")
-        );
-    }
-
-    #[cfg(all(feature = "rt-tokio", feature = "http-body-0-4-x"))]
-    #[tokio::test]
-    async fn path_based_bytestreams() -> Result<(), Box<dyn std::error::Error>> {
-        use super::ByteStream;
-        use bytes::Buf;
-        use std::io::Write;
-        use tempfile::NamedTempFile;
-        let mut file = NamedTempFile::new()?;
-
-        for i in 0..10000 {
-            writeln!(file, "Brian was here. Briefly. {}", i)?;
-        }
-        let body = ByteStream::from_path_body_0_4(&file).await?.into_inner();
-        // assert that a valid size hint is immediately ready
-        assert_eq!(body.content_length(), Some(298890));
-        let mut body1 = body.try_clone().expect("retryable bodies are cloneable");
-        // read a little bit from one of the clones
-        let some_data = body1
-            .next()
-            .await
-            .expect("should have some data")
-            .expect("read should not fail");
-        assert!(!some_data.is_empty());
-        // make some more clones
-        let body2 = body.try_clone().expect("retryable bodies are cloneable");
-        let body3 = body.try_clone().expect("retryable bodies are cloneable");
-        let body2 = ByteStream::new(body2).collect().await?.into_bytes();
-        let body3 = ByteStream::new(body3).collect().await?.into_bytes();
-        assert_eq!(body2, body3);
-        assert!(body2.starts_with(b"Brian was here."));
-        assert!(body2.ends_with(b"9999\n"));
-        assert_eq!(body2.len(), 298890);
-
-        assert_eq!(
-            ByteStream::new(body1).collect().await?.remaining(),
-            298890 - some_data.len()
-        );
-
-        Ok(())
     }
 
     #[cfg(feature = "rt-tokio")]
